+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - Removed the `accessibilityLabel` prop from `<LineChart/>` and `<BarChart/>` in favour of a more complete accessibility approach
 - `<LineChart/>` now requires `xAxisLabels`
 - `<Sparkline />` now accepts a `series` prop instead of `data` and universal styling props, so that multiple data series can be displayed and individually configured
+- Removed the accessibilityLabel from `<NormalizedStackedBarChart/>`, component is accessible by default.
 
 ### Added